refactor more in transform part of api
