+++ src/core/selector/selector.ts
@@ -1,0 +1,16 @@
+export type Selector =
+'A' |
+'B' |
+'C' |
+'D' |
+'E' |
+'F' |
+'G' |
+'H' |
+'I' |
+'J' |
+'O' |
+'M' |
+'N' |
+'0' |
+'1'
